add miniconf
