+++ Makefile
@@ -1,4 +1,4 @@
-PYTHON_FILES = main.py scripts/
+PYTHON_FILES = main.py miniconf/ scripts/
 JS_FILES = $(shell find static/js -name "*.js")
 CSS_FILES = $(shell find static/css -name "*.css")
 .PHONY: format-python format-web format run freeze format-check
